Internal: Enhances variable manager with renew promotions [ED-21889]
linking to the upgrade page.

It also includes a disabled prop for the label field.

Relates to ED-21889
