+++ packages/packages/core/editor-variables/src/components/fields/label-field.tsx
@@ -35,6 +35,7 @@ export type LabelFieldProps = {
 	showWarningInfotip?: boolean;
 	variables?: TVariablesList;
 	onKeyDown?: ( event: KeyboardEvent< HTMLInputElement > ) => void;
+	disabled?: boolean;
 };
 
 export const LabelField = ( {
@@ -49,6 +50,7 @@ export const LabelField = ( {
 	showWarningInfotip = false,
 	variables,
 	onKeyDown,
+	disabled,
 }: LabelFieldProps ) => {
 	const [ label, setLabel ] = useState( value );
 	const [ errorMessage, setErrorMessage ] = useState( '' );
@@ -90,6 +92,7 @@ export const LabelField = ( {
 			} }
 			// eslint-disable-next-line jsx-a11y/no-autofocus
 			autoFocus={ focusOnShow }
+			disabled={ disabled }
 		/>
 	);
